Arrange tests and schema.sql
