+++ books.json
@@ -1,0 +1,1 @@
+[{"title":"Book1","publisher":"Virag","cover_state":"good","publish_date":"2020/09/08"},{"title":"Book 2","publisher":"Micronaut","cover_state":"bad","publish_date":"2003/04/30"},{"title":"Awesome Book","publisher":"Anna","cover_state":"good","publish_date":"1980-02-23"},{"title":"Book123","publisher":"Bob","cover_state":"bad","publish_date":"2005/11/12"}]
