+++ schema.sql
@@ -1,34 +1,29 @@
 CREATE DATABASE catalog;
 USE catalog;
 
+CREATE TABLE book(id INT GENERATED BY DEFAULT AS IDENTITY, 
+publisher VARCHAR, cover_state VARCHAR, archived BOOLEAN);
+
+CREATE TABLE label(id INT GENERATED BY DEFAULT AS IDENTITY, title VARCHAR, color VARCHAR, PRIMARY KEY (id))
+
+CREATE TABLE genre (id INT NOT NULL AUTO_INCREMENT, name VARCHAR(30),items VARCHAR(30));
+
+CREATE TABLE musicAlbum(id INT NOT NULL AUTO_INCREMENT, on_spotify BOOLEAN, publish_date DATE, archived boolean);
+
+CREATE TABLE author(id INT GENERATED BY DEFAULT AS IDENTITY, first_name VARCHAR, last_name VARCHAR);
+
+CREATE TABLE game(id INT GENERATED BY DEFAULT AS IDENTITY, multiplayer VARCHAR, last_played DATE, archived BOOLEAN)
+
+
 CREATE TABLE item(id INT GENERATED BY DEFAULT AS IDENTITY, 
 publish_date DATE, 
 archived BOOLEAN,
- books_id INT, 
- genre INT NOT NULL,
- PRIMARY KEY (id),
-  FOREIGN KEY (books_id)
-   REFERENCES books(id));
-  foreign key(genre) REFERENCES genre(id),
+label_id INT, 
+author_id INT,
+genre_id INT NOT NULL,
+PRIMARY KEY (id),
+FOREIGN KEY (label_id)
+REFERENCES label(id),
+FOREIGN KEY (genre) REFERENCES genre(id)
+FOREIGN KEY (author_id) REFERENCES author(id));
 
-CREATE TABLE books(id INT GENERATED BY DEFAULT AS IDENTITY, 
-publisher VARCHAR,
- cover_state VARCHAR);
-
-
-
-
-
-
-CREATE TABLE genre (
-	id INT NOT NULL AUTO_INCREMENT,
-    name VARCHAR(30),
-    items VARCHAR(30)
-);
-
-CREATE TABLE musicAlbum(
-   id INT NOT NULL AUTO_INCREMENT,
-   on_spotify BOOLEAN,
-   publish_date DATE,
-   archived boolean       
-);
